+++ ResistorColorCodeCalculator.Logic/IOhmValueCalculator.cs
@@ -14,6 +14,6 @@ namespace ResistorColorCodeCalculator.Logic
         /// <param name="bandBColor">The color of the second significant figure band.</param>
         /// <param name="bandCColor">The color of the decimal multiplier band.</param>
         /// <param name="bandDColor">The color of the tolerance value band.</param>
-        int CalculatorOhmValue(string bandAColor, string bandBColor, string bandCColor, string bandDColor);
+        decimal CalculatorOhmValue(string bandAColor, string bandBColor, string bandCColor, string bandDColor);
     }
 }
